requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
spacy>=3.6.0
pymysql>=1.1.0
cryptography>=41.0.0
dnspython>=2.4.0
idna>=3.4